+++ Examples/cwd/main.swift
@@ -1,0 +1,4 @@
+import SystemExtras
+import SystemPackage
+
+try print(FilePath.workingDirectory())