+++ CHANGELOG.md
@@ -2,3 +2,5 @@
 
 - Add `FileMetadata` type to reperesent file metadata
 - Add `FilePath.metadata()` method to retrieve file metadata
+- Add `FilePath.workingDirectory()` static method for returning current working directory.
+- Add `FilePath.setWorkingDirectory()` static method for setting current working directory.